backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9  # PostgreSQL adapter (optional, for PostgreSQL support)

# Validation and Security
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator>=2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0

# AI and LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-aws==0.1.0
openai>=1.0.0
sentence-transformers==2.2.2

# Vector Database
chromadb==0.4.18

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
reportlab==4.0.7
pillow>=10.0.0

# Audio Processing
pydub==0.25.1
moviepy==1.0.3
openai-whisper
gtts==2.5.1

# File Handling
aiofiles==23.2.1

# HTTP Client
httpx==0.25.2

# Email
fastapi-mail==1.4.1
jinja2>=3.1.2  # For email templates

# AWS
boto3>=1.34.72

